update backend to handle zip files and treat them as potential shapefiles; parse these and produce the same json as when using kml; move writedatatoexcel to exceplapi.py
